+++ src/auth.ts
@@ -1,4 +1,4 @@
-import * as clc from "cli-color";
+import * as clc from "colorette";
 import * as FormData from "form-data";
 import * as fs from "fs";
 import * as http from "http";
@@ -19,7 +19,6 @@ import * as scopes from "./scopes";
 import { clearCredentials } from "./defaultCredentials";
 import { v4 as uuidv4 } from "uuid";
 import { randomBytes, createHash } from "crypto";
-import { bold } from "cli-color";
 import { track } from "./track";
 import {
   authOrigin,
@@ -32,7 +31,6 @@ import {
   googleOrigin,
 } from "./api";
 
-/* eslint-disable camelcase */
 // The wire protocol for an access token returned by Google.
 // When we actually refresh from the server we should always have
 // these optional fields, but when a user passes --token we may
@@ -83,7 +81,6 @@ interface GitHubAuthResponse {
   scope: string;
   token_type: string;
 }
-/* eslint-enable camelcase */
 
 // Typescript emulates modules, which have constant exports. We can
 // overcome this by casting to any
@@ -471,7 +468,7 @@ async function loginRemotely(): Promise<UserCredentials> {
   logger.info();
   logger.info("1. Take note of your session ID:");
   logger.info();
-  logger.info(`   ${bold(sessionId.substring(0, 5).toUpperCase())}`);
+  logger.info(`   ${clc.bold(sessionId.substring(0, 5).toUpperCase())}`);
   logger.info();
   logger.info("2. Visit the URL below on any device and follow the instructions to get your code:");
   logger.info();
@@ -576,7 +573,7 @@ async function loginWithLocalhost<ResultType>(
     server.listen(port, () => {
       logger.info();
       logger.info("Visit this URL on this device to log in:");
-      logger.info(clc.bold.underline(authUrl));
+      logger.info(clc.bold(clc.underline(authUrl)));
       logger.info();
       logger.info("Waiting for authentication...");