+++ src/commands/hosting-disable.ts
@@ -1,4 +1,4 @@
-import * as clc from "cli-color";
+import * as clc from "colorette";
 
 import { Client } from "../apiv2";
 import { Command } from "../command";