+++ src/emulator/controller.ts
@@ -1,9 +1,9 @@
-import * as clc from "cli-color";
+import * as clc from "colorette";
 import * as fs from "fs";
 import * as path from "path";
 
 import { logger } from "../logger";
-import { track } from "../track";
+import { track, trackEmulator } from "../track";
 import * as utils from "../utils";
 import { EmulatorRegistry } from "./registry";
 import {
@@ -14,13 +14,14 @@ import {
   EMULATORS_SUPPORTED_BY_UI,
   isEmulator,
 } from "./types";
-import { Constants, FIND_AVAILBLE_PORT_BY_DEFAULT } from "./constants";
+import { Constants, FIND_AVAILBLE_PORT_BY_DEFAULT, DEFAULT_HOST, DEFAULT_PORTS } from "./constants";
 import { EmulatableBackend, FunctionsEmulator } from "./functionsEmulator";
 import { parseRuntimeVersion } from "./functionsEmulatorUtils";
 import { AuthEmulator } from "./auth";
 import { DatabaseEmulator, DatabaseEmulatorArgs } from "./databaseEmulator";
 import { FirestoreEmulator, FirestoreEmulatorArgs } from "./firestoreEmulator";
 import { HostingEmulator } from "./hostingEmulator";
+import { EventarcEmulator } from "./eventarcEmulator";
 import { FirebaseError } from "../error";
 import { getProjectId, needProjectId, getAliases, needProjectNumber } from "../projectUtils";
 import { PubsubEmulator } from "./pubsubEmulator";
@@ -34,7 +35,11 @@ import { EmulatorLogger } from "./emulatorLogger";
 import * as portUtils from "./portUtils";
 import { EmulatorHubClient } from "./hubClient";
 import { promptOnce } from "../prompt";
-import { FLAG_EXPORT_ON_EXIT_NAME, JAVA_DEPRECATION_WARNING } from "./commandUtils";
+import {
+  FLAG_EXPORT_ON_EXIT_NAME,
+  JAVA_DEPRECATION_WARNING,
+  MIN_SUPPORTED_JAVA_MAJOR_VERSION,
+} from "./commandUtils";
 import { fileExistsSync } from "../fsutils";
 import { StorageEmulator } from "./storage";
 import { getStorageRulesConfig } from "./storage/rules/config";
@@ -134,19 +139,6 @@ async function getAndCheckAddress(emulator: Emulators, options: Options): Promis
 }
 
 /**
- * Starts a specific emulator instance
- * @param instance
- */
-export async function startEmulator(instance: EmulatorInstance): Promise<void> {
-  const name = instance.getName();
-
-  // Log the command for analytics
-  void track("Emulator Run", name);
-
-  await EmulatorRegistry.start(instance);
-}
-
-/**
  * Exports emulator data on clean exit (SIGINT or process end)
  * @param options
  */
@@ -158,7 +150,7 @@ export async function exportOnExit(options: any) {
         `Automatically exporting data using ${FLAG_EXPORT_ON_EXIT_NAME} "${exportOnExitDir}" ` +
           "please wait for the export to finish..."
       );
-      await exportEmulatorData(exportOnExitDir, options);
+      await exportEmulatorData(exportOnExitDir, options, /* initiatedBy= */ "exit");
     } catch (e: any) {
       utils.logWarning(e);
       utils.logWarning(`Automatic export to "${exportOnExitDir}" failed, going to exit now...`);
@@ -355,9 +347,8 @@ export async function startAll(
       `No emulators to start, run ${clc.bold("firebase init emulators")} to get started.`
     );
   }
-  const deprecationNotices: string[] = [];
   if (targets.some(requiresJava)) {
-    if (!(await commandUtils.checkJavaSupported())) {
+    if ((await commandUtils.checkJavaMajorVersion()) < MIN_SUPPORTED_JAVA_MAJOR_VERSION) {
       utils.logLabeledError("emulators", JAVA_DEPRECATION_WARNING, "warn");
       throw new FirebaseError(JAVA_DEPRECATION_WARNING);
     }
@@ -366,7 +357,8 @@ export async function startAll(
   hubLogger.logLabeled("BULLET", "emulators", `Starting emulators: ${targets.join(", ")}`);
 
   const projectId: string = getProjectId(options) || ""; // TODO: Next breaking change, consider making this fall back to demo project.
-  if (Constants.isDemoProject(projectId)) {
+  const isDemoProject = Constants.isDemoProject(projectId);
+  if (isDemoProject) {
     hubLogger.logLabeled(
       "BULLET",
       "emulators",
@@ -410,6 +402,19 @@ export async function startAll(
     }
   }
 
+  function startEmulator(instance: EmulatorInstance): Promise<void> {
+    const name = instance.getName();
+
+    // Log the command for analytics
+    void track("Emulator Run", name);
+    void trackEmulator("emulator_run", {
+      emulator_name: name,
+      is_demo_project: String(isDemoProject),
+    });
+
+    return EmulatorRegistry.start(instance);
+  }
+
   if (shouldStart(options, Emulators.HUB)) {
     const hubAddr = await getAndCheckAddress(Emulators.HUB, options);
     const hub = new EmulatorHub({ projectId, ...hubAddr });
@@ -432,6 +437,10 @@ export async function startAll(
     const foundMetadata = findExportMetadata(importDir);
     if (foundMetadata) {
       exportMetadata = foundMetadata;
+      void trackEmulator("emulator_import", {
+        initiated_by: "start",
+        emulator_name: Emulators.HUB,
+      });
     } else {
       hubLogger.logLabeled(
         "WARN",
@@ -466,7 +475,7 @@ export async function startAll(
   }
 
   if (shouldStart(options, Emulators.EXTENSIONS)) {
-    const projectNumber = Constants.isDemoProject(projectId)
+    const projectNumber = isDemoProject
       ? Constants.FAKE_PROJECT_NUMBER
       : await needProjectNumber(options);
     const aliases = getAliases(options, projectId);
@@ -532,6 +541,20 @@ export async function startAll(
     });
     await startEmulator(functionsEmulator);
   }
+  // Since the functions emulator will always be started by default, and the eventarc emulator
+  // is only used by the functions emulator, we will start them in conjunction (using the default host/port)
+  // even if the app developer has not explicitly configured the eventarc emulator.
+  if (!shouldStart(options, Emulators.EVENTARC)) {
+    if (options.config.src.emulators) {
+      options.config.src.emulators.eventarc = { host: DEFAULT_HOST, port: DEFAULT_PORTS.eventarc };
+    }
+  }
+  const eventarcAddr = await getAndCheckAddress(Emulators.EVENTARC, options);
+  const eventarcEmulator = new EventarcEmulator({
+    host: eventarcAddr.host,
+    port: eventarcAddr.port,
+  });
+  await startEmulator(eventarcEmulator);
 
   if (shouldStart(options, Emulators.FIRESTORE)) {
     const firestoreLogger = EmulatorLogger.forEmulator(Emulators.FIRESTORE);
@@ -558,6 +581,10 @@ export async function startAll(
         `Importing data from ${exportMetadataFilePath}`
       );
       args.seed_from_export = exportMetadataFilePath;
+      void trackEmulator("emulator_import", {
+        initiated_by: "start",
+        emulator_name: Emulators.FIRESTORE,
+      });
     }
 
     const config = options.config;
@@ -657,6 +684,11 @@ export async function startAll(
       const databaseExportDir = path.resolve(importDirAbsPath, exportMetadata.database.path);
 
       const files = fs.readdirSync(databaseExportDir).filter((f) => f.endsWith(".json"));
+      void trackEmulator("emulator_import", {
+        initiated_by: "start",
+        emulator_name: Emulators.DATABASE,
+        count: files.length,
+      });
       for (const f of files) {
         const fPath = path.join(databaseExportDir, f);
         const ns = path.basename(f, ".json");
@@ -687,7 +719,7 @@ export async function startAll(
       const importDirAbsPath = path.resolve(options.import);
       const authExportDir = path.resolve(importDirAbsPath, exportMetadata.auth.path);
 
-      await authEmulator.importData(authExportDir, projectId);
+      await authEmulator.importData(authExportDir, projectId, { initiatedBy: "start" });
     }
   }
 
@@ -723,7 +755,7 @@ export async function startAll(
       utils.assertIsString(options.import);
       const importDirAbsPath = path.resolve(options.import);
       const storageExportDir = path.resolve(importDirAbsPath, exportMetadata.storage.path);
-      storageEmulator.storageLayer.import(storageExportDir);
+      storageEmulator.storageLayer.import(storageExportDir, { initiatedBy: "start" });
     }
   }
 
@@ -770,15 +802,25 @@ export async function startAll(
     await startEmulator(ui);
   }
 
+  let serviceEmulatorCount = 0;
   const running = EmulatorRegistry.listRunning();
   for (const name of running) {
     const instance = EmulatorRegistry.get(name);
     if (instance) {
       await instance.connect();
     }
+    if (ALL_SERVICE_EMULATORS.includes(name)) {
+      serviceEmulatorCount++;
+    }
   }
 
-  return { deprecationNotices };
+  void trackEmulator("emulators_started", {
+    count: serviceEmulatorCount,
+    count_all: running.length,
+    is_demo_project: String(isDemoProject),
+  });
+
+  return { deprecationNotices: [] };
 }
 
 /**
@@ -786,7 +828,7 @@ export async function startAll(
  * @param exportPath
  * @param options
  */
-export async function exportEmulatorData(exportPath: string, options: any) {
+export async function exportEmulatorData(exportPath: string, options: any, initiatedBy: string) {
   const projectId = options.project;
   if (!projectId) {
     throw new FirebaseError(
@@ -847,7 +889,7 @@ export async function exportEmulatorData(exportPath: string, options: any) {
 
   utils.logBullet(`Exporting data to: ${exportAbsPath}`);
   try {
-    await hubClient.postExport(exportAbsPath);
+    await hubClient.postExport({ path: exportAbsPath, initiatedBy });
   } catch (e: any) {
     throw new FirebaseError("Export request failed, see emulator logs for more information.", {
       exit: 1,