+++ src/emulator/storage/files.ts
@@ -18,6 +18,8 @@ import { RulesetOperationMethod } from "./rules/types";
 import { AdminCredentialValidator, FirebaseRulesValidator } from "./rules/utils";
 import { Persistence } from "./persistence";
 import { Upload, UploadStatus } from "./upload";
+import { trackEmulator } from "../../track";
+import { Emulators } from "../types";
 
 interface BucketsList {
   buckets: {
@@ -108,6 +110,9 @@ export type CopyObjectRequest = {
   incomingMetadata?: IncomingMetadata;
   authorization?: string;
 };
+
+// Matches any number of "/" at the end of a string.
+const TRAILING_SLASHES_PATTERN = /\/+$/;
 
 export class StorageLayer {
   constructor(
@@ -285,6 +290,7 @@ export class StorageLayer {
       throw new Error(`Unexpected upload status encountered: ${upload.status}.`);
     }
 
+    const storedMetadata = this.getMetadata(upload.bucketId, upload.objectId);
     const filePath = this.path(upload.bucketId, upload.objectId);
     const metadata = new StoredFileMetadata(
       {
@@ -306,7 +312,10 @@ export class StorageLayer {
       ["b", upload.bucketId, "o", upload.objectId].join("/"),
       upload.bucketId,
       RulesetOperationMethod.CREATE,
-      { after: metadata?.asRulesResource() },
+      {
+        before: storedMetadata?.asRulesResource(),
+        after: metadata?.asRulesResource(),
+      },
       upload.authorization
     );
     if (!authorized) {
@@ -391,12 +400,15 @@ export class StorageLayer {
    */
   public async listObjects(request: ListObjectsRequest): Promise<ListObjectsResponse> {
     const { bucketId, prefix, delimiter, pageToken, authorization } = request;
+
     const authorized = await this._rulesValidator.validate(
-      ["b", bucketId, "o", prefix].join("/"),
+      // Firebase Rules expects the path without trailing slashes.
+      ["b", bucketId, "o", prefix.replace(TRAILING_SLASHES_PATTERN, "")].join("/"),
       bucketId,
       RulesetOperationMethod.LIST,
       {},
-      authorization
+      authorization,
+      delimiter
     );
     if (!authorized) {
       throw new ForbiddenError();
@@ -418,10 +430,13 @@ export class StorageLayer {
       if (delimiter) {
         const delimiterIdx = name.indexOf(delimiter);
         const delimiterAfterPrefixIdx = name.indexOf(delimiter, prefix.length);
-        // items[] contains object metadata for objects whose names do not contain delimiter, or whose names only have instances of delimiter in their prefix.
+        // items[] contains object metadata for objects whose names do not contain
+        // delimiter, or whose names only have instances of delimiter in their prefix.
         includeMetadata = delimiterIdx === -1 || delimiterAfterPrefixIdx === -1;
         if (delimiterAfterPrefixIdx !== -1) {
-          // prefixes[] contains truncated object names for objects whose names contain delimiter after any prefix. Object names are truncated beyond the first applicable instance of the delimiter.
+          // prefixes[] contains truncated object names for objects whose names contain
+          // delimiter after any prefix. Object names are truncated beyond the first
+          // applicable instance of the delimiter.
           prefixes.add(name.slice(0, delimiterAfterPrefixIdx + delimiter.length));
         }
       }
@@ -504,7 +519,7 @@ export class StorageLayer {
    * Export is implemented using async operations so that it does not block
    * the hub when invoked.
    */
-  async export(storageExportPath: string) {
+  async export(storageExportPath: string, options: { initiatedBy: string }): Promise<void> {
     // Export a list of all known bucket IDs, which can be used to reconstruct
     // the bucket metadata.
     const bucketsList: BucketsList = {
@@ -513,6 +528,11 @@ export class StorageLayer {
     for (const b of await this.listBuckets()) {
       bucketsList.buckets.push({ id: b.id });
     }
+    void trackEmulator("emulator_export", {
+      initiated_by: options.initiatedBy,
+      emulator_name: Emulators.STORAGE,
+      count: bucketsList.buckets.length,
+    });
     // Resulting path is platform-specific, e.g. foo%5Cbar on Windows, foo%2Fbar on Linux
     // after URI encoding. Similarly for metadata paths below.
     const bucketsFilePath = path.join(storageExportPath, "buckets.json");
@@ -544,10 +564,16 @@ export class StorageLayer {
    * Import can be implemented using sync operations because the emulator should
    * not be handling any other requests during import.
    */
-  import(storageExportPath: string) {
+  import(storageExportPath: string, options: { initiatedBy: string }): void {
     // Restore list of buckets
     const bucketsFile = path.join(storageExportPath, "buckets.json");
     const bucketsList = JSON.parse(readFileSync(bucketsFile, "utf-8")) as BucketsList;
+    void trackEmulator("emulator_import", {
+      initiated_by: options.initiatedBy,
+      emulator_name: Emulators.STORAGE,
+      count: bucketsList.buckets.length,
+    });
+
     for (const b of bucketsList.buckets) {
       const bucketMetadata = new CloudStorageBucketMetadata(b.id);
       this._buckets.set(b.id, bucketMetadata);
@@ -555,6 +581,14 @@ export class StorageLayer {
 
     const metadataDir = path.join(storageExportPath, "metadata");
     const blobsDir = path.join(storageExportPath, "blobs");
+
+    // Handle case where export contained empty metadata or blobs
+    if (!existsSync(metadataDir) || !existsSync(blobsDir)) {
+      logger.warn(
+        `Could not find metadata directory at "${metadataDir}" and/or blobs directory at "${blobsDir}".`
+      );
+      return;
+    }
 
     // Restore all metadata
     const metadataList = this.walkDirSync(metadataDir);