+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-- Adds warnings about upcoming deprecation of `--token`, `FIREBASE_TOKEN`, and `login:ci`.
-- Fix ENAMETOOLONG bug when uploading files with long paths in storage emulator. (#4470)
+- Fixes an issue where pushing an empty buffer from the android sdks crashes the storage emulator (#4843)
+- Adds emulation support for extensions that emit events / custom event triggers (#4464)
+- Fixes issue with batchCreate in auth emulator in which it errored for empty mfaInfo (#4848)
+- Fixes scroll bars in the Firestore Emulator UI